New algorithm + functions
I've changed the algorithm for sorting through the unique words and keep track of how many times they appear in the TomSawyer textfile.
The algorithm now no longer saves all the words appearing in the TomSawyer textfile but instead reads a word and saves it in a string variable to determine whether it is unique and keep track of its count. It will then save the word in the struct if it is unique and move on to the next word. Each time the string variable is allocated for the next word and the entire textfile data is not stored. I've also added two new functions for the algorithm which help me determine whether I am observing a unique word for the first time or not. The functions are checker and loca.