Added the fifth and sixth functions
The fifth function transmits a message through the nodes until it reaches the country you wanted to the message to be sent to. It keeps track of how many times a message passed through the node as well as what the message is. The sixth function prints the path of the nodes in the list (the order).